added draggable edges for barriers

+++ trunk/simulations-java/simulations/waveinterference/src/edu/colorado/phet/waveinterference/model/SlitPotential.java
@@ -114,7 +114,7 @@ public class SlitPotential implements Potential {
         update();
     }
 
-    public double getSlitWidth() {
+    public int getSlitWidth() {
         return slitWidth;
     }
